refactor: /tools/socket - 2 files instead of 1
