+++ src/services/messages/:messageId/put.ts
@@ -7,10 +7,10 @@ import Channel from '../../../models/Channel'
 import Member from '../../../models/Member'
 import Message from '../../../models/Message'
 import { commonErrorsMessages } from '../../../tools/config/constants'
-import { emitToMembers } from '../../../tools/socket/socket'
 import { BadRequestError } from '../../../tools/errors/BadRequestError'
 import { ForbiddenError } from '../../../tools/errors/ForbiddenError'
 import { NotFoundError } from '../../../tools/errors/NotFoundError'
+import { emitToMembers } from '../../../tools/socket/emitEvents'
 
 export const putByIdMessagesRouter = Router()
