+++ src/tools/socket/index.ts
@@ -1,0 +1,15 @@
+import { Server as HttpServer } from 'http'
+import { Server as HttpsServer } from 'https'
+import { Server as SocketIoServer } from 'socket.io'
+
+interface Socket {
+  io: null | SocketIoServer
+  init: (httpServer: HttpServer | HttpsServer) => void
+}
+
+export const socket: Socket = {
+  io: null,
+  init (httpServer) {
+    socket.io = new SocketIoServer(httpServer)
+  }
+}
